requirements.txt
streamlit
ultralytics
pillow
requests
wikipedia-api
torch
torchvision
opencv-python-headless
numpy
